+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Logpoint 2 - 2023-09-29
 ### Language
+- Make binding names case-sensitive
 - Add `^` syntax to terminate modifier parsing. There is a basic example [in the tutorial](http://uiua.org/docs/functions#terminating-modifiers).
 - Add shadowable [constants](https://uiua.org/docs/constants) like `e`.
 - Change [trident`∋`](https://uiua.org/docs/trident) argument order to make it easier to reason about